refactor: split paths.rs (434 lines) into paths/ module directory
